Move beautifier to submodule
